+++ yuyo/__init__.py
@@ -50,6 +50,7 @@ __all__: list[str] = [
     "InteractionError",
     "ModalClient",
     "ModalContext",
+    "Page",
     "ReactionClient",
     "ReactionHandler",
     "ReactionPaginator",
@@ -99,6 +100,7 @@ from .list_status import ServiceManager
 from .list_status import TopGGService
 from .modals import ModalClient
 from .modals import ModalContext
+from .pagination import Page
 from .pagination import aenumerate
 from .pagination import async_paginate_string
 from .pagination import paginate_string