+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [yuyo.pagination.Page][] is now exported top-level at `Yuyo.Page`.
+
 ## [1.17.0] - 2023-08-18
 ### Added
 - Wait for implementation for modals.