+++ CHANGELOG.md
@@ -8,6 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - [yuyo.pagination.Page][] is now exported top-level at `Yuyo.Page`.
 
+### Fixed
+- Properly track method names for callbacks registered using
+  [as_interactive_button][yuyo.components.as_interactive_button].
+  This allows the custom ID metadata to be overridden  by passing
+  `{"function_name": "METADATA"}` to
+  [ActionColumnExecutor.\_\_init\_\_][yuyo.components.ActionColumnExecutor.__init__].
+
 ## [1.17.0] - 2023-08-18
 ### Added
 - Wait for implementation for modals.