+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.18.0] - 2023-08-23
 ### Added
 - [yuyo.components.builder][] (a class descriptor),
   [ActionColumnExecutor.add_builder][yuyo.components.ActionColumnExecutor.add_builder], and
@@ -801,7 +802,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Iffy behaviour around "locking" the reaction executor which lead to some requests just being ignored.
 
-[Unreleased]: https://github.com/FasterSpeeding/Yuyo/compare/v1.17.0...HEAD
+[Unreleased]: https://github.com/FasterSpeeding/Yuyo/compare/v1.18.0...HEAD
+[1.18.0]: https://github.com/FasterSpeeding/Yuyo/compare/v1.17.0...v1.18.0
 [1.17.0]: https://github.com/FasterSpeeding/Yuyo/compare/v1.16.0...v1.17.0
 [1.16.0]: https://github.com/FasterSpeeding/Yuyo/compare/v1.15.0...v1.16.0
 [1.15.0]: https://github.com/FasterSpeeding/Yuyo/compare/v1.14.2...v1.15.0